Added Problem 8a and 8b Solutions
Summary
=======

The Day 8 problems focused on a very basic assembly language, using 3
instructions (acc, jmp, & nop) and an accumulator global variable. I
will not spoil the problems any further.

Notes
=====

I kind of thought of creating an environment dictionary (see [Peter
Norvig's Lisp in Python](https://norvig.com/lispy.html) tutorial) with the 3 instructions as the key and their functions as the value. Along with these functions would be an accumulator global variable.

However, due to the small size of this "programming language" and how I felt it would bloat the program, I decided to not to and just stick with some basic if
statements.